A few fruit, and MIMIC_KNOW
Adds a few fruit - these are uncapitalized which will be the rule for
all other new items.
They include the choke-apple, and the MIMIC_KNOW flag to support it
(items with this flag look like another item until identified).
Currently the mimiced item is hard coded to the previous one, but
that will change.
W