Spicy stores, initial commit
Includes reanimation of CHR, introduction of Speed-as-a-Stat, the ability
to save and restore RNG state to a memory buffer, and support for doing
more interesting thing with stores.

There is at least partial code to do some of these things, including
stealing, getting new stock, reshuffling, buying and selling the store and
fighting.