Speed stat potion added
At this point "speed as a stat" should work but no attempt has been made
to balance it.

The 'simple' way (where player-calc just adds speed stat to Speed, adjusted so
 10 is +0) has the upshot that you can get up to +18 to speed intrinsically
(for a race/class with +0), but not more than +30 total permanent (although
you can exceed this with timed effects).

The effect of which is that characters shortly after stat-gain are much more
powerful, late-game characters may be a little more powerful, speed penalties
(including weight) matter more, and so does timed speed. Loadouts start to
change when you have enough speed items to hit +30, with permanent speed
being less important.

This is not all bad and adds up to something that is balancable as-is (by
shifting stat potion depths etc.) but might still work better if more changes
were made (there is nothing magic about maxed stat = +18 or the +30 limit, it
could go through a table)