Metric magic number to fix AC/weight power
This is not just a metric change, though - it now uses to_a as well as ac
and as this means high +s are now priced higher, the overall scale is reduced.
The driver for this was that low base-AC items weren't getting recognized as
good items - a lighter [0,+12] was considered cheaper than a heavier [6,+3].
This made sense when enchantment was easy, but it isn't now.