Make more money fast
The cash items have been replaced. This made for less nicely
distributed types (there are a lot more types now) so money_kind
was modified to give a more varied distribution, while make_gold
was modified in the hope of making cash drops more interesting -
they scale up faster and there are more unusually large drops.
This is especially true in no-selling, where the previous flat *5
was replaced with a higher chance of large drops (with roughly the
same average effect, although it's depth dependent).
The clip to max short (because pvals are 16 bit) was replaced with
a retry loop, as this is now common enough that $32767 items would
be obvious. (For the same reasons - if pvals do become 32-bit,
don't just rip out this loop without changing the constants, or
you'll be an instant zillionaire)