Blow 'em up
Allow stores to be destroyed permanently.
This means a flag added to the store (set to blow up the store
next time you return to town), an added member per grid as a
general purpose tag (to store the store ID per grid) and the
code to replace all grids with a given tag with rubble.