A new, fairer method of rolling HP
The intent is to make any character of the same hitdie have the same
HP at level 50 - the same as the average result of the previous method
(full hitdie for level 1, randint(1, hitdie) for all others) - and HP
close to the average at all other levels (never more than 5% below).

It does this by first rolling 49 random values, then rerolling a random
one of these and taking the maximum or minimum of the original and new
values, to increase or decrease the sum. Once it has a set of 49 rolls
which sum to the target, it starts swapping them to reduce midlevel
deviations.

This swapping involves scoring the values by calculating the sum of
squares of negative deviations, swapping a randomly selected pair of
values, calculating the new sum of squares, and keeping the version
with the lower sum of squares. This process stops when no level between
5 and 40 has a HP below the mean by more than 5%.