Quests, and the Rats Quest
Sufficient support for quests has been added to handle the Rats
quest - ask at the General Store, and a new stair will be created
from the town. The level there is a vault of a new category, with
quest specific monsters. Information on quests and the current
state of completion can be had from the knowledge menu. Kill all
the rats and return, and you will get a reward - these are hard
coded rather than being defined in quest.txt as there are so many
possibilities for different rewards to different character types,
different ways to complete a quest, etc. You will also get other
quests (currently the incomplete 'Cave' quest) unlocked.