Abilities: parser
An ability is a player flag - entirely positive, entirely negative or a
mixture and includes "talents" which can be bought with talent points at
birth or later, and "mutations" which are acquired through a mutation - and
it may include some abilities which are neither, or both.

They are stored as player race/class flags and tested in the same way, but
have an additional ability structure. These are added to ability[] when the
ability.txt is parsed - a player flag with ability[<flag>] == NULL is not
an ability.