Large store/home support
Correct overflow (wrong letters for store with items off the end of the list
of directly selectable items), make the help more helpful, increase the
default maximum size but add an override for this in store.txt and save it
per-store in the savefile rather than relying only on initialization.

This is intended primarily as support for a home which can get (much) bigger
during the course of the game, e.g. as a reward for completing a task.