Closable stores
Stores have a persistent 'open' flag and entrance position.
This allows them to be closed and opened (removing and adding the entrance
grid) at any time. They are open by default, but it is possible for them
to be closed initially from store.txt.

This is intended to support task rewards etc. which open, close or replace
stores (or your home).