Printers / pack activation
Printer activation is mostly complete (although there is as yet no way
to specify that an object is unprintable other than not having a material,
or that it requires special conditions other than having the needed chunks
and printer for its material).

This includes the ability to activate items from the pack. Printers have this
as they can't be wielded - lights because it's clunky to wield something that
you didn't intend to keep wielded.