Tweak effects of hunger
There are now no bad effects from only being "Hungry" rather than "Weak".
Partly because the effects were not obvious and so players would assume
that being Hungry was harmless - and partly because the penalty to +dam
especially was way too harsh. (e.g. a level 1 character with good physical
state and a mundane Main Gauche was reduced to 0 damage without being Weak)
Also, this means that there is now no hidden dependency on the relationship
between Hungry and Weak (assuming that it is roughly 2:1 through the lack
variable, scaled 0 to 20, being compared to 10).

Similar penalties to skills still exist - they started at a point close to
Weak before and still do. The penalty to hit has been nerfed but is still
significant (-1 when just Weak increasing to -7 when near starving). The
penalty to damage has been removed entirely because of its potential to
screw up wimpy characters, but has been replaced by "real" weakness: -1
to Strength when barely Weak increasing to -11 when starving. This was
done with stat_add, which works but might not be ideal: it might be best
to signal this in a way that is clearly different, e.g. orange highlight
on the left panel and character sheet.