Rip out the spellcasting system
You can no longer use m or p to cast (although the command remains -
it will be used for the new system). You can also not use G to gain
(the new system will not use that). Player SP has also been removed
along with effects and items referring to it.

Monster spells, magic devices and the element of mana are untouched.

The new spell system is not included in this patch with the exception
of a few changes in wording ("intrinsic abilities" - unfortunately
an overload of the word 'ability'...) but is expected to consist of
spell-like abilities gained from race, class, abilities, mutations,
worn equipment etc. with the limiting that SP provides coming from
a combination of cooldowns and HP use - and not requiring a book or
other item to cast.