New casting 3
There is now a (inaccurately named) class_magic structure in the player
race, shapechange, object kind, ego item and artifact structures. These
are combined to give the list of available abilities. Shapechange is the
exception - if you are shapechanged, all other abilities become
unavailable. Parsers have been added (and there are renames of some of
the members to prevent conflicts with lines used by other files).