Rip out books
Books are now gone, as tvals and objects, along with most references
to them as the new casting system doesn't require them. They do exist
as abstract named lists of techniques - this is not normally visible
but is still around as it could be used to classify techniques, e.g.
to use a two-stage selection (book, then technique) if you had too
many techniques to fit on screen in one page.