Blows added from an ability
Melee blows can be added from an a ability. These are 'free blows',
taking no energy, taken at the end of every round of melee combat.
They are defined by a verb ("butt"), random value for damage and
optional element. They use the same to-hit calculations as standard
combat (without a weapon), and apply immunity or vulnerability to
the element if used.