Random effects on abilities, + Horns
An ability can have a effect specified with a random delay. Every
game turn there is a 1 in <delay> chance of the effect chain being
called.

This is used for Horns - this will at random times honk (waking
monsters, as aggravate monster). It tries to be smart, though -
being much more likely to honk if dangerous monsters are nearby
(especially if in LOS), and asleep.