Every item is special, every item is great...
All artifacts are now generated by a method similar to that used
for special artifacts. Previously, make_artifact would usually fail
(success rate strongly level dependent) - it now always succeeds,
if there are any artifacts left (or any artifacts of that tval, if
one is specified). It is called for a proportion of objects which
increases with level, and is increased if a good or great item was
requested. The new make_artifact selects artifacts randomly based
on their rarity, scaled down when away from their native level.

The advantages of this model are that artifact probabilities are
decoupled from their base items' probabilities (which made some
artifacts have unexpected probability spikes because their base
item was of a level far from the artifact's), that it is easier
to change the way that artifacts are distributed now that this
comes from a single function rather than the interaction with
base item distribution, and that the order of artifacts in the
file no longer affects the distribution (as it did when the file
was scanned from beginning to end for a match - an earlier one
succeeding blocked later ones from generating).

Known limitations are that the distribution of artifacts produced
per level is still strongly influenced by the numbers, levels and
types of items and monsters created in the level - and that a
sufficiently patient grinder could sit on level 1 until all the
artifacts were generated (when only high level artifacts are left,
that is what would be produced). That though would be its own
punishment IMO.

A new mode for the wizard stats generation has been added to show
all artifacts' probabilities by level generated at.