Extension races
An extension is similar to a player race - it uses the same structure
and is loaded from the same file. It is intended to modify one or more
races rather than standing alone, adding a variation such as "Cyborg",
"Mutant" etc. The p_race.txt file has been extended to allow for
declaring an extension (with "ext" instead of "name") and defining
which extensions are allowed for each race in the "exts" lines.
Allowing no extensions is OK, the default and used for Androids.

Also included is the Mutant extension (p_mutant.c) which is allowed
for Humans and Time-Lords. It reduces stats a little, grants a few
immediate (and a few revealed as you advance) non-nasty mutations and
allows an increasing chance with level of fighting off a mutation or
the loss of a mutation (i.e. you get to reroll a different mutation).
At high levels you get resistance to radiation.