Ego generation, part 1
This cleans up the init code and rearranges ego selection so that
(like with artifacts) the first step is to decide based on level
and good/great whether to generate an ego, and if so an ego type
is then selected (without having first produced an object kind).
The kind is then selected from the possible kinds for that ego,
weighted by the level's item generation probabilities.

This gets rid of problems caused by base item rarity influencing
ego generation, although further tuning is needed - some problems
appear to be caused by the hard cut off at max depth, the stepwise
OOD roll or the linear goodness chance which are unmodified by this
change, in roughly reducing order of importance (although they do
now become much easier to modify independently).