Ego generation, part 2
Modify ego item selection so that ego probabilities peak at their minimum
depth, reducing linearly to zero at the maximum depth + 1 (instead of the
previous constant probability over the full range). This smooths out the
spikes caused when a common ego or egos are past their maximum depth and
suddenly other competing egos become more common (which are smaller but
still present with ego item types being selected before object kinds - it
is competing with all other egos wather than just the ones for that item)

When out of depth, a calculation is used that is similar in prniciple to
that used before - dividing by (levels out of depth / 3) - but it now
changes the probabilities directly, avoiding the integer rounding and
separate one-ine-n roll that caused "stair-stepping" (jumps every 3
levels).