Ego generation, part 3
Replace the ego generation chance per level.
This needs to be done, since now a successful ego roll means "definitely
produce an ego" not "maybe produce an ego if the kind is right". It is also
an opportunity to rebalance ego chances, although this will need to be
revisited when the ego and item lists are closer to complete.

Normal items now have a chance of becoming an ego item of 3.25% at level 0
increasing linearly to 10% below level 40 (down from 10% at level 0
increasing to 30% at level 66).
Good items have a fixed chance of 15% (down from 30%).
Great items have a chance of 80% at level 0 /reducing/ to 20% at level 95+,
down from 100%. The reasoning for the reverse dependency is that a great
item should be something the character really wants. At level 1, almost
any ego item is a big upgrade - while late on you would rather have endgame
consumables than another pair of speed boots.

The results of these changes are that effective probabilities as measured
with diving sims are within 10% of original ones on levels 5, 10, 20 and 40.
Below that, it becomes less generous (producing around half as many egos on
level 95).