Bones, part 1
Read bones files, by rewriting "Randy, the Random" (as in store random
uniques). Functions for creating, comparing and format conversion of
bones files headers.

Bones files are primarily parsable text, resembling one entry from
monster.txt. There's a binary header for compatible backwardness, and
the format conversion crud is there so it is independent of byte order
and sizeof(long). (The text is of course naturally CPU independent, but
if they are ever to be handled by an unmodified Hearse client they need
to superficially resemble Nethack bones.)