Clean up experience
The experience is stored as an integer and separate fractional component.
By rewriting the experience-on-monster-death code, it is now easier to
drop in changes to how experience is calculated (for example, to make it
dependent on the number already killed, or to give partial credit for
reducing a monster's HP without killing it). As the fractional calculations
are now confined to one function, the monster-info code (lore_append_exp)
can use the new code as well, removing the previous duplication.

This does not however itself change the experience calculation (except
perhaps for changing the effect of rounding errors!), or the meaning of
the exp and exp_frac fields.