Experience soft cap
A early Experience pill can be a cheasy 'skip the first half of the game'.
In some variants it's a problem with monsters as well (and it may be that
some of Xygos's quests are cheasable).

So experience is now reduced if it requires gaining more than one level at
once. The first level gain is free (as you might have been very close to
it already), but the remaining experience after the second level is halved.
This repeats so that if there is still enough experience to gain a third
level the experience left is quartered, etc.

The result is that a level-1 Experience pill can give 8 levels, down to a
few in the mid game - enough to feel like a big boost, but not too much.
There is nothing magical about the number 2 (which it divides by each
level) though, so the effect could easily be tweaked in either direction.