Airports
Towns have an airport, which can be used to buy tickets to other towns.
Prices and flight times vary depending on their distances - to limit the
amount of town-hopping that can be done (searching for stuff to buy) the
prices are stiff and there is only one flight to each destination per
day, so compulsive travellers will soon start to see nasties in town.