Separate stores
Each town now has a separate array of stores.
The original stores pointer is still used, and most of store.c and
ui-store.c still uses it - it always now points to the player's current
town's stores array. (The current town is the most recent one, even you
are current in the dungeon.)

All are saved and loaded, maintained daily and searched for objects.
Checking for store destruction is still only when the player is there
(which is OK just as it before, when it could be deferred until the
player next visited town).

At game startup, all towns have all stores that aren't marked closed.
They do get randomized in the usual way to select owner and stock.
Anything such as a quest that changes store properties will have to be
aware of multitowns if it wants to affect stores outside the current
town, though.