Recalculate player->class as needed
The player->class pointer now points to the "primary class" - the one
used for all things that require a single class. It is calculated
from the lev_class[] array at birth, loading and level change, as the
class with the most levels in it. (If a tie-breaker is needed, the one
with the earliest first level wins.)