Renamed to Xygos 0.0.1
This is supposed to allow Xygos and V to be installed simultaneously
without stepping on each other, and to always tell the user in visible
version numbers, help files etc. that this is Xygos 0.0.1 (an Angband
variant), not Angband 4.2.2. It doesn't change internal variables,
comments etc., pref files, developer oriented documentation etc.
The main HTML manual is also unchanged for now.