Add hidden activations
The HIDE_ACTIVATION flag is for situations such as batteries.
These have an activation (Androids can eat them) but displaying
this is misleading as non-Androids can't eat them and even for
Androids, you eat rather than activate them so it is still wrong.

This simply removes all activation messages for all items with
this flag (currently only batteries).