Fix for object_fully_known bug
There was a situation (seen with lights and cybergear) in which
object_learn_unknown_icon() was unable to find any new icons to
learn, but object_fully_known() wouldn't admit this - resulting
in the "Learn flavor, any effect and all the icons" loop in store.c
becoming infinite.

This change to object_non_fault_icons_known prevents that from
occuring.