Fix "startup delay and too many charges in shops" bug
The store_maint() loop was trying to create more items than could
be produced. This was fixed in the edit files by allowing some
random items, but the loop was also modified to make this less
likely to reoccur (if it is unable to create enough items it will
fall back to smaller numbers of items).