Giant Form improvements
More consistent use of check_shapechanged(). Eating is the exception,
because you can eat from the floor. A message is given to tell you
that (as otherwise you have been told that you have nothing to eat
when you have a pack full of food) which looks less wrong. The
messages you get when you try to enter a store when changed now have
a special case for the home.