Vault verify and new tvals
New characters are added (= was ring, is now printer/block. " was
amulet, is now cyberware). Vaults now have characters verified as they
are loaded, ensuring they are acceptable. Currently this assumes that
all [A-Z][a-z] besides Xx is a monster, though. Some vaults that
failed this check have been modified (replacing rings and amulets).