Carried flags
Carried flags are object-flags that take effect when an object is
carried (and not equipped). They are displayed on the character sheet
in a combined column between equipment and player columns. So far
only one object uses them (the cheese). They could be added to ego
items and artifacts as well, though there is as yet no parser for them.