Generalize player flags
Give objects (and kinds, egos, artifacts) a set of player flags which are
combined with the other sources of player flags when you are wearing the
equipment.

Add a link between player flags and timed effects, so that every player
flag can also be granted temporarily as a timed effect.

Modify the talents UI to display equipment and temporary abilities in colour.

Add "Stone Skin" mutation, and use it to replace the old "STONESKIN" flag.

This does not so far handle forbidden combinations of flags (for example,
it doesn't prevent you from wearing a armor of Snake Skin when you have
intrinsic Steel Skin). There is also nothing to stop players (think
Androids) who would normally be unable to access mutations at all from
obtaining them through equipment or temporary buffs, although this isn't
necessarily a Bad Thing.