Fix shop initialization again
Shops were requiring a lot of repeated object generations to get
sufficient stock, with most being merged into existing similar
items. The result was slow startup and large stacks of items.
This removes some unnecessary looping, bumps up the maximum level
and good/great chance when lots of objects have beem called for,
as well as not always carrying the extra items that are produced.