Extend generalized player flags
This handles forbids and requires between abilities, now that they can
be equipment or temporary effects. The intrinsic abilities now only
consider other intrinsic abiltiies (not equipment or temporary) - this
means that you can't use a temporary effect as a stepping stone to
gain a permanent effect with prerequisites that you don't have
innately.

Temporary and equipment abilities that are forbidden by intrinsic,
temporary or equipment abilities don't take effect. Requirements
between temporary and equipment abilities have not been implemented,
on the basis that an item granting an effect should be expected to
always do so, without having any surprising hidden requirements.