Multiclass: changing class at level up
When levelling up, if a class switch has been asked for then change class.
Your exp will be reduced by one level, but the resulting level may go up
or down as you will have a different mix of exp multipliers. If it goes
down, your max level will not be changed (and so your level but not
experience will appear drained).