Multiclass: abilities
Abilities with a class and level as prerequisite (e.g. Marksman's
specializations) take the level from the best level in class, that
you have at least 1 level in and which is in the list of classes
for the ability. Additionally, you need to have the right primary
class (one in the list of classes for the ability).

Other player flags are unchanged - meaning that they are taken
from the primary class.