Enhanced character sheet resists mode
There were some problems with this mode - the list of slots was wide enough
that it didn't fit, and one list was long enough that it didn't fit
vertically. So the top-left and top-right areas were removed (as they are
identical to the other mode, nothing has been lost). The two middle
columns have been swapped (so that the long list wasn't sharing a column
with the sustains block). While it now fits in 80x24, it will expand to fit a
wider terminal if present. This has involved making it more generic - the
number of columns can change (although so far it does not), the width is
per column, the column used for sustain could be changed, etc.