Cybersalon has a faction
It bases prices mostly on your rank - which can be advanced by selling
non-storebought cyberware to them. The quality and quantity of items
is dependent on rank, and until you become a member there will be no
items for sale. The price for nonmembers is bad enough that you can
expect to do better selling to the Black Market - in order to give
you the choice between more cash, or more future possibilities.