Random names for store owners
These are taken from two lists, of first and second names in
files first.txt and second.txt which are originally from
Mika Spåra's OXP (Oolite expansion pack), "Random Player/
Ship Names".

Fixed names can still be used - and are for the airport and
HQ. If a * is used for a name then it is replaced with a random
one when stores are reset.