Wrestlers get combat bonuses
Wrestlers if unarmed and not too heavily armored get a level-in-class
dependent + to blows, hit and damage. You can also get critical hits,
which increase damage and sometimes stun, slow or hold. Higher-level
and less-armored wrestlers get more (with heavy armor you will be
effectively at 1/5 level), while high-level and high-AC monsters are
harder to get crits against.

For the hold status to be useful, it can't be cancelled on the next
blow. So the "attacking breaks holding" rule has been removed. This
makes all forms of holding more powerful, which may be undesirable.
The best way to deal with this might be to distinguish paralysis
(which is broken by a hit, like sleep) from holding (which cannot,
but might be broken if you move away).