Multiple alternative effects
Any effect list may include one or more NEXT entries. Calling
effect_do() takes a parameter, which is the number of NEXT blocks
to be skipped - the result is that this parameter can select between
different sets of effects. By default this is 0, meaning the first
(or only) set. The sets are distinguished in the object info.

One intended use (mostly not included here) is the Clown's "Rare
Card" technique - a class power that allows a Clown to use cards
in a different (and more powerful) way. It might also be used for
Engineer's class powers for devices, etc.