Ego and other data changes
Sort death messages, remove or rename some egos, hints and lies,
names, a more thematic win message and the ability to prepend ego
item names in the same way as artifacts, using a leading '<'.

Add a debug command to generate one of every ego, and allow start
items to be generated even when you are full (they spill onto the
floor).