Mutants - create, load, limit level etc, radiation effects
Mutants may be randomly created, with a probability which can be set per monster
but which defaults to a z_info constant (1 in 25).

Load/save mutants.

Parse and make use of minimum and maximum monster race levels to use per mutation -
this is on top of the level check (which is there so that mutants are allocated
at appropriate levels, similarly to base monsters, not to limit the places that
a mutation is found).

Sets the 'mutant'/'modified' flags and check whether a base monster is an
appropriate target - it must not have existing flags that the mutation would add,
not be unique/special, not be radiation resistant etc.

Add special effects from a radiation hit - monsters may grow, split, polymorph or
mutate. (Growth is like polymorph with a predefined target. Splitting only
happens if they can already multiply.)

Also sort and add a death message, and allow radiation balls from wonder.