Randomizable quest artifacts
Randomizable quest artifacts. These are present independently of
random-artifact mode - they are always randomized but have a fixed
power (defined in artifacts.txt), name and description. They are
intended as rewards for midbosses and dungeon guardians, and don't
generate randomly. Also fixes some bugs, and removes the need to
add a kind in object.txt for quest artifacts (fixed or random).