Airport does not destroy
The airport is needed to get to the Fortress town (to kill the 4 first
win bosses) and then to get to orbit (for the last one). So there must
always be a way to get to Fortress town (before the 4 bosses are gone)
and to orbit (afterwards).

This change prevents the airport from every being destroyed which is
not ideal but will work until there is a replacement.