Add multiego forbids
Forbids are lines such as forbid:fireproof|acid, of which there can
be at most one per ego item (but that line can have any number of
forbidden items - and as each one is a case insensitive substring
match, it is possible to match classes of egos e.g. forbid:* will
match any ego with a * in it.)

A multiego can't be generated which contains a pair of forbidden
egos - the direction doesn't matter. (So putting forbid:fireproof
under the frostproof ego, and putting forbid:frostproof under the
fireproof ego will have the same effect of ensuring that no item
with both frost- and fireproof egos can be generated.)