Branding / Mundanity cards
The formatting card turns any piece of equipment into a base item,
removing artifact, ego, enhancement, faults etc. Uses - to deal
with nasty items with a useful base item, to prep an item for
branding when it has an unwanted ego and as Schmuck Bait.

The transformation card is its opposite: it tries to add an ego.
Artifacts and maxed-out multi-egos always fail. Otherwise, non-ego
items normally succeed and apply a random ego. Sometimes (1 in 4)
a second ego can be applied to an ego item. Occasionally (1 in 25
for both) it will instead act as a formatting card, or apply two
egos to a non-ego item.
The power of all these ego items is based on the player's maximum
depth.