The Dark Helmet
The Dark Helmet is the boss of the Volcano.
Includes a stormtrooper monster, with a new CONFUSED flag so the
stormtroopers shoot each other as much as you.
Includes S_FRIENDS, which summons a new batch of the friends it was
generated with.
Includes an artifact, with changes to obj-desc (so that an artifact with
first character '!' is described only by the artifact's name: "the Dark
Helmet", not "the dark helmet "The Dark Helmet").
The artifact makes use of RAND_ flags, which were previously only used
on ego-items - any object (including special artifacts) can now use them.
The handling of the flags has been fixed and extended so that
combinations of flags can be used on the same item, and have been made a
bit more random in the process: the RES_POWER now gives 1/6 high resists,
1/3 low resists and 1/2 powers, and there is a small chance of any RAND_
flag applying more than once.