Cyber Salon doesn't restock items sold
The Salon gives faction from items sold to it.
While they are checked for not originating in the store, this doesn't
prevent an item originally found in the dungeon from being sold and
bought again any number of times, accumulating unlimited faction.
So simply don't restock sold items in the Salon.