Fix Cyborgs' net connection talent
Cyborgs get the Net Connection talent for free at birth. They were
being prompted at the talent screen for it, because the extension
initialization (which granted the talent) hadn't run yet. This fix
moves init around so that by the time talents are prompted for, the
talent has been added. The tourist init is modified as this means
that it is being given $ before, not after start items are processed.
(The tourist has start items worth more than typical starting $,
so it matters.)