Stealing from stores, and town faction
Stealing is updated with a difficulty level based on DEX, CHR and factions.
There's also a help text and precentages of outcomes displayed.
Town faction is used for chance of successully stealing, level and number of
items generated in stores.
Faction (BM and town) will return to zero after a long time spent without
stealing (or doing anything else that reduces it below 0). Cyber Salon's
faction doesn't because you can sell it items to increase faction.