Remove/rename/replace the last of the original equipment
Cloaks, Bucklers, Gloves are gone, replaced with part of a new set.
This includes a fix for always lines in store.txt with no sval, and
allows item lines in ego_item.txt with a leading ! to be used to
remove an item from the list allowed for that ego.