Engineer item techniques
This gives several item related techniques to Engineers:

Diagnose
Identify one fault on an item.

Troubleshoot
Attempt to remove a fault from an item. If you are successful, you get
to fling it at enemies as a lag bolt.

Tweak
Improves equipment, improving the to-hit and to-dam bonuses of weapons
and the to-AC of armour.

Reverse Engineer
Identify one unknown icon of any kind on an item.

Upgrade
Attempts to make a non-ego into an ego item. On a non-ego base item it
will almost always succeed, although the results can't be predicted.
Rarely, it will add multiple egos. It can be run against an ego item in
order to create a multiple ego, but this doesn't often work and risks
destroying it.

Inject Fault
Attempt to add a fault to an item. (A few faults can be useful in their
own right, but the real reason to use this is Troubleshoot's lag bolts.)

Service
Attempt to remove all faults from an item, with greater chance of
success than Troubleshoot (like the difference between major-repair and
repair cards)

Recycle
Converts a suitable item (made of a printable material and not too small
or large) back into blocks of raw material suitable for printing.

Rebuild
Attempt to make a fragile item as robust as new again.

Format
Attempts to remove all additions, positive and negative, from a piece of
equipment - to-hit and -dam, egos, activations, faults and so on. Even
artifacts will lose everything which makes them special, with no volcano
needed.