New monsters, old monsters rearranged and supporting monster flags
EVASIVE: Blows randomly miss. Unlike AC, it's not dependent on to-hit
but only on speed difference. It doesn't have any effect on ranged
weaopns. It's for things such as birds that should be especially
difficult to hit but not soak up much damage and should be used
sparingly - it could get annoying if every fruit bat had it.

RAND_100: All moves are random.

SUCK: An attack method. Similar to CRAWL, but more thematic for slugs,
leeches etc.

AQUATIC: For fish etc. which can't leave water (or possibly lava)