Xorns -> Aliens, IM_EDGED, IM_BLUNT
The X character is now for nonhumanoid aliens which include but are
not limited to xorn/xarens.
The IM_EDGED and IM_BLUNT flags have been added so that monsters
(currently the xorn and xaren are the only ones, both IM_EDGED) can
resist edged or blunt weapons. In original melee mode (O-combat is
similar but more complicated) damage caused by damage dice or + to
damage (rather than slays, brands, strength bonuses...) is reduced
to 1/4 if you fight with the arong type of weapon. (Shooting is not
affected.)