Add some bugs
Including IN_LEVEL for the pill bug, which stops it from becoming
less common below its native depth. It's there because pill bugs
are never a threat but exist for the side effects (negative - losing
pills - and positive, gaining them.) It could be used for any other
monster that fits that description.