Fix colour cycling, improve verification
Monster color cycling was being ignored because at the time that
visuals_cycler_set_cycle_for_race() was called, ridx was unset (still
zero for all monsters), resulting in the visuals_color_cycles_by_race
being wrong.
This change stores the cycle information in a temporary list, allowing
the visuals_cycler_set_cycle_for_race() calls to be done later when
ridx is known.
It also prevents crashing through reading an uninitialized pointer if
an unrecognized color cycle name is given, and prints a warning when
one is found.

Also includes some unrelated fixes to verification of the monster list.