Feeling threshold autobalancer
THIS IS WIP as it will not converge to a monotonically increasing table,
even after 23K iters.

This has been added because the object and monster lists change and the
hard coded thresholds gave bad results, and simply changing these
thresholds would soon be obsoleted. So instead, a method of automatically
deriving thresholds is used which can be run again as needed.

This adds a stats mode (Ctrl-A ~) which creates levels and extracts the
object and monster ratings, which have been changed from int to double.
The 9 feelings are separated by 8 thresholds, and each threshold is
different for each depth. The methods of calculating ratings have also
changed. The object value is no longer squared and the bonus for being
out of depth has been reduced because values already increase quickly
with depth, making this too swingy. The monster value is now less
dependent on being out of depth - previously a level 91 monster
encountered at level 90 would be worth twice as much as if it were seen
at level 90, for example. Exceptions are also made for uniques (which
are worth more) and pit and vault monsters (which are worth less, as you
can usually avoid them - and they tended to dominate the ratings).

It runs a series of iterations, each one being a new set of randarts and
a new level for each depth. For each depth, ratings from each iteration
are sorted, with the thresholds taken from a set of points indexing that
array (for example, after 100 iterations the threshold between level 1
and 2 would be at index 15 and the succeeding ones at 35, 55, 70, 80, 93,
and 98 so that for any number of iterations, there would always be 15%
of levels of feeling 1, 35% of feeling 1 or 2, etc.)

Each iteration extracts a set (one per level/threshold) of these values,
and every 10 iterations each level/threshold's set of values is
individually sorted and the midpoint (i.e. the median) extracted. The
results are written to a file (feel.txt) in a suitable format to paste
into generate.c. If a key is pressed, results are also output to the
console before returning (it will run indefinitely if no key is pressed).

To give an idea of progress, a line describing the number of "bad" entries,
such as "Iter 9: bad obj 465 / 365, bad mon 482 / 366, sum 1678" is output.
The first of each pair is the count of entries greater than the next depth,
the second is the count of entries greater than the next higher threshold.
They are all measures of noise remaining in the table and so should
decrease over time.

Expect it to take a while - using -O3 -flto is recommended. It also appears
to be a good way to discover rare bugs...