In-town quests
Adds two quests that take place in town.
This involves artifact start items (for the reward), and a way to specify
that a quest is in the town with a lake. Water resistance is added.
Also a few rewords etc. including a message when day/night changes after
a flight.